Scene 파일 나누기
파일을 신 별로 나누고 폴더를 만듬
