+++ tetris/tetris/GamePlayScene.cpp
@@ -1,0 +1,10 @@
+#include "Scene.h"
+#include "GamePlayScene.h"
+#include "stdafx.h"
+
+GamePlayScene::GamePlayScene() {}
+GamePlayScene::~GamePlayScene() {}
+
+void GamePlayScene::draw_main() { //게임판 그리는 함수
+
+}
