+++ tetris/tetris/Scene.cpp
@@ -11,12 +11,3 @@ Scene::~Scene() {
 		delete m_pGameClient;
 }
 
-WaitScene::WaitScene() {}
-WaitScene::~WaitScene() {}
-
-PlayScene::PlayScene() {}
-PlayScene::~PlayScene() {}
-
-void PlayScene::draw_main() { //게임판 그리는 함수
-    
-}
