+++ tetris/tetris/Scene.h
@@ -20,25 +20,3 @@ protected:
     GameClient* m_pGameClient;
 };
 
-
-class WaitScene : public Scene {
-
-    WaitScene();
-    ~WaitScene();
-
-    void update();
-};
-
-class PlayScene : public Scene {
-
-    PlayScene();
-    ~PlayScene();
-
-    void update();
-    void reset();
-    void reset_main();
-    void reset_main_cpy();
-    void draw_map();
-    void new_block();
-    void draw_main();
-};
