+++ tetris/tetris/WaitScene.cpp
@@ -1,0 +1,6 @@
+#include "Scene.h"
+#include "Waitscene.h"
+#include "stdafx.h"
+
+WaitScene::WaitScene() {}
+WaitScene::~WaitScene() {}
