+++ tetris/tetris/WaitScene.h
@@ -1,0 +1,11 @@
+#pragma once
+
+class Scene;
+
+class WaitScene : public Scene {
+
+    WaitScene();
+    ~WaitScene();
+
+    void update();
+};
